+++ ios/pusher_client/Sources/pusher_client/ConnectionListener.swift
@@ -5,7 +5,8 @@
 //  Created by Romario Chinloy on 10/27/20.
 //
 
-import PusherSwiftWithEncryption
+import Foundation
+import PusherSwift
 
 class ConnectionListener: PusherDelegate {
     static let `default`: ConnectionListener = ConnectionListener()